Add AppConfig that handles the URL for api host and raw base url.
