+++ lib/AppConfig.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:meta/meta.dart';
+
+class AppConfig extends InheritedWidget {
+  AppConfig({
+    @required this.apiEndpoint,
+    @required this.host,
+    @required Widget child,
+  }) : super(child: child);
+
+  final String apiEndpoint;
+  final String host;
+
+  static AppConfig of(BuildContext context) {
+    return context.inheritFromWidgetOfExactType(AppConfig);
+  }
+
+  @override
+  bool updateShouldNotify(InheritedWidget oldWidget) => false;
+}
