+++ lib/components/Menu.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import './MultiContent.dart';
 
 class Menu extends StatelessWidget {
   Widget menuItem(String menuTitle) {
