+++ test/widget_test.dart
@@ -4,10 +4,7 @@
 // find child widgets in the widget tree, read text, and verify that the values of widget properties
 // are correct.
 
-import 'package:flutter/material.dart';
 //import 'package:flutter_test/flutter_test.dart';
-
-import 'package:feed_flutter_ge/main.dart';
 
 void main() {
 //  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
